@@ -1,161 +1,26 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Microsoft.Azure.CognitiveServices.Vision.Face;
-using Microsoft.Azure.CognitiveServices.Vision.Face.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace FaceApp
 {
     public class Program
-    {       
-        
-
+    {
         public static void Main(string[] args)
         {
-            BuildWebHost(args).Run();
-             //Authenticate.
-           
-
-            const string IMAGE_BASE_URL = "https://csdx.blob.core.windows.net/resources/Face/Images/";
-            const string RECOGNITION_MODEL4 = RecognitionModel.Recognition04;
-            //// Detect - get features from faces.
-            //DetectFaceExtract(client, IMAGE_BASE_URL, RECOGNITION_MODEL4).Wait();
-            //// Find Similar - find a similar face from a list of faces.
-            //FindSimilar(client, IMAGE_BASE_URL, RECOGNITION_MODEL4).Wait();
-            //// Verify - compare two images if the same person or not.
-            //Verify(client, IMAGE_BASE_URL, RECOGNITION_MODEL4).Wait();
-
-            //// Identify - recognize a face(s) in a person group (a person group is created in this example).
-            //IdentifyInPersonGroup(client, IMAGE_BASE_URL, RECOGNITION_MODEL4).Wait();
-            //// LargePersonGroup - create, then get data.
-            //LargePersonGroup(client, IMAGE_BASE_URL, RECOGNITION_MODEL4).Wait();
-            //// Group faces - automatically group similar faces.
-            //Group(client, IMAGE_BASE_URL, RECOGNITION_MODEL4).Wait();
-            //// FaceList - create a face list, then get data
+            CreateHostBuilder(args).Build().Run();
         }
 
-        public static IFaceClient Authenticate(string endpoint, string key)
-        {
-            return new FaceClient(new ApiKeyServiceClientCredentials(key)) { Endpoint = endpoint };
-        }
-
-        public static async Task DetectFaceExtract(IFaceClient client, string url, string recognitionModel)
-        {
-            Console.WriteLine("========DETECT FACES========");
-            Console.WriteLine();
-
-            // Create a list of images
-            List<string> imageFileNames = new List<string>
-                            {
-                                "detection1.jpg",    // single female with glasses
-								// "detection2.jpg", // (optional: single man)
-								// "detection3.jpg", // (optional: single male construction worker)
-								// "detection4.jpg", // (optional: 3 people at cafe, 1 is blurred)
-								"detection5.jpg",    // family, woman child man
-								"detection6.jpg"     // elderly couple, male female
-							};
-
-            foreach (var imageFileName in imageFileNames)
-            {
-                IList<DetectedFace> detectedFaces;
-                
-                // Detect faces with all attributes from image url.
-                detectedFaces = await client.Face.DetectWithUrlAsync($"{url}{imageFileName}",
-                        returnFaceAttributes: new List<FaceAttributeType> { FaceAttributeType.Accessories, FaceAttributeType.Age,
-                        FaceAttributeType.Blur, FaceAttributeType.Emotion, FaceAttributeType.Exposure, FaceAttributeType.FacialHair,
-                        FaceAttributeType.Gender, FaceAttributeType.Glasses, FaceAttributeType.Hair, FaceAttributeType.HeadPose,
-                        FaceAttributeType.Makeup, FaceAttributeType.Noise, FaceAttributeType.Occlusion, FaceAttributeType.Smile },
-                        // We specify detection model 1 because we are retrieving attributes.
-                        detectionModel: DetectionModel.Detection01,
-                        recognitionModel: recognitionModel);
-
-                Console.WriteLine($"{detectedFaces.Count} face(s) detected from image `{imageFileName}`.");
-                // </snippet_detect>
-                // <snippet_detect_parse>
-                // Parse and print all attributes of each detected face.
-                foreach (var face in detectedFaces)
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
                 {
-                    Console.WriteLine($"Face attributes for {imageFileName}:");
-
-                    // Get bounding box of the faces
-                    Console.WriteLine($"Rectangle(Left/Top/Width/Height) : {face.FaceRectangle.Left} {face.FaceRectangle.Top} {face.FaceRectangle.Width} {face.FaceRectangle.Height}");
-
-                    // Get accessories of the faces
-                    List<Accessory> accessoriesList = (List<Accessory>)face.FaceAttributes.Accessories;
-                    int count = face.FaceAttributes.Accessories.Count;
-                    string accessory; string[] accessoryArray = new string[count];
-                    if (count == 0) { accessory = "NoAccessories"; }
-                    else
-                    {
-                        for (int i = 0; i < count; ++i) { accessoryArray[i] = accessoriesList[i].Type.ToString(); }
-                        accessory = string.Join(",", accessoryArray);
-                    }
-                    Console.WriteLine($"Accessories : {accessory}");
-
-                    // Get face other attributes
-                    Console.WriteLine($"Age : {face.FaceAttributes.Age}");
-                    Console.WriteLine($"Blur : {face.FaceAttributes.Blur.BlurLevel}");
-
-                    // Get emotion on the face
-                    string emotionType = string.Empty;
-                    double emotionValue = 0.0;
-                    Emotion emotion = face.FaceAttributes.Emotion;
-                    if (emotion.Anger > emotionValue) { emotionValue = emotion.Anger; emotionType = "Anger"; }
-                    if (emotion.Contempt > emotionValue) { emotionValue = emotion.Contempt; emotionType = "Contempt"; }
-                    if (emotion.Disgust > emotionValue) { emotionValue = emotion.Disgust; emotionType = "Disgust"; }
-                    if (emotion.Fear > emotionValue) { emotionValue = emotion.Fear; emotionType = "Fear"; }
-                    if (emotion.Happiness > emotionValue) { emotionValue = emotion.Happiness; emotionType = "Happiness"; }
-                    if (emotion.Neutral > emotionValue) { emotionValue = emotion.Neutral; emotionType = "Neutral"; }
-                    if (emotion.Sadness > emotionValue) { emotionValue = emotion.Sadness; emotionType = "Sadness"; }
-                    if (emotion.Surprise > emotionValue) { emotionType = "Surprise"; }
-                    Console.WriteLine($"Emotion : {emotionType}");
-
-                    // Get more face attributes
-                    Console.WriteLine($"Exposure : {face.FaceAttributes.Exposure.ExposureLevel}");
-                    Console.WriteLine($"FacialHair : {string.Format("{0}", face.FaceAttributes.FacialHair.Moustache + face.FaceAttributes.FacialHair.Beard + face.FaceAttributes.FacialHair.Sideburns > 0 ? "Yes" : "No")}");
-                    Console.WriteLine($"Gender : {face.FaceAttributes.Gender}");
-                    Console.WriteLine($"Glasses : {face.FaceAttributes.Glasses}");
-
-                    // Get hair color
-                    Hair hair = face.FaceAttributes.Hair;
-                    string color = null;
-                    if (hair.HairColor.Count == 0) { if (hair.Invisible) { color = "Invisible"; } else { color = "Bald"; } }
-                    HairColorType returnColor = HairColorType.Unknown;
-                    double maxConfidence = 0.0f;
-                    foreach (HairColor hairColor in hair.HairColor)
-                    {
-                        if (hairColor.Confidence <= maxConfidence) { continue; }
-                        maxConfidence = hairColor.Confidence; returnColor = hairColor.Color; color = returnColor.ToString();
-                    }
-                    Console.WriteLine($"Hair : {color}");
-
-                    // Get more attributes
-                    Console.WriteLine($"HeadPose : {string.Format("Pitch: {0}, Roll: {1}, Yaw: {2}", Math.Round(face.FaceAttributes.HeadPose.Pitch, 2), Math.Round(face.FaceAttributes.HeadPose.Roll, 2), Math.Round(face.FaceAttributes.HeadPose.Yaw, 2))}");
-                    Console.WriteLine($"Makeup : {string.Format("{0}", (face.FaceAttributes.Makeup.EyeMakeup || face.FaceAttributes.Makeup.LipMakeup) ? "Yes" : "No")}");
-                    Console.WriteLine($"Noise : {face.FaceAttributes.Noise.NoiseLevel}");
-                    Console.WriteLine($"Occlusion : {string.Format("EyeOccluded: {0}", face.FaceAttributes.Occlusion.EyeOccluded ? "Yes" : "No")} " +
-                        $" {string.Format("ForeheadOccluded: {0}", face.FaceAttributes.Occlusion.ForeheadOccluded ? "Yes" : "No")}   {string.Format("MouthOccluded: {0}", face.FaceAttributes.Occlusion.MouthOccluded ? "Yes" : "No")}");
-                    Console.WriteLine($"Smile : {face.FaceAttributes.Smile}");
-                    Console.WriteLine();
-                }
-            }
-        }
-
-
-        public static IWebHost BuildWebHost(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseApplicationInsights()
-                .UseStartup<Startup>()
-                //.ConfigureLogging(builder => {
-                //    builder.AddLog4Net("log4net.config");
-                //})
-                .Build();
-
+                    webBuilder.UseStartup<Startup>();
+                });
     }
 }
