@@ -1,262 +1,8 @@
-﻿@model FaceApp.Models.Face
-@{
-    ViewData["Title"] = "Identify";
-    Layout = "~/Views/Shared/_Layout.cshtml";  
-    
+﻿@{
+    ViewData["Title"] = "Home Page Change";
 }
 
-<div>
-    <form asp-action="Index" id="Identifyform" enctype="multipart/form-data">
-
-        <div class="content text-center">
-            <h5 class="mob-heading1">
-                Attendance
-            </h5>
-            <div class="capture-wrap" style="position:relative">
-                <img src="~/images/camera-capture.png" class="img-responsive mx-auto">
-                <video id="stream"
-                       style="position: absolute;
-                                left: 0;
-                                right: 0;
-                                margin: auto;
-                                top: 0;
-                                bottom: 0;
-                                border-radius:46px;"></video>
-                <canvas id="capture" width="320" height="240"></canvas>                
-                    @if (ViewBag.imgcap != null)
-                    {
-                        int width = 338;
-                        string height = "height=378";
-                        if(ViewBag.browerName != null)
-                        {
-                            if(ViewBag.browerName == "true")
-                            {
-                                width = 352;
-                                height = "";
-                            }
-                        }
-                        <input id="attendanceImg" width="@width" @height type="image" disabled="disabled" src="@ViewBag.imgcap" style="position: absolute;
-                                left: 0;
-                                right: 0;
-                                margin: auto;
-                                top: 0;
-                                bottom: 0;
-                                border-radius:46px;" />                               
-                    }
-                    else
-                    {
-                        <input type="image" disabled="disabled" id="imgcap" name="imgcap" style="position: absolute;
-                                left: 0;
-                                right: 0;
-                                margin: auto;
-                                top: 0;
-                                bottom: 0;
-                                border-radius:46px;display:none" />
-                    }
-                    <input type="text" id="txtimgcap" style="display:none" asp-for="CapturedFile" />  
-                    <input type="text" id="browerName" name="browerName" style="display:none"/>
-            </div>
-            <div class="text-center p20">
-                @*<a href="" id="btn-capture" class="btn btn-lg btn-primary">Take Attendance</a>*@
-                @if (ViewBag.Result != null)
-                {
-                    @ViewBag.Result
-
-                }
-                else
-                {
-                    <button id="btn-capture" type="submit" class="btn btn-success" style="margin-bottom:8px;"> Take Attendance</button>
-                }
-
-            </div>
-        </div>
-
-        @*<div class="col-md-12">
-            <div class="col-md-4">
-                <h2>Identify a Person</h2>
-            </div>
-        </div>
-        <div class="col-md-6">
-            <div class="panel panel-primary">
-                <div class="panel-heading">Camera</div>
-        <div class="panel-body" style="text-align:center">
-                <div id="my_camera"></div>
-                A button for taking snaps
-        <video id="stream" width="320" height="240"></video>
-                <button id="btn-capture" type="submit" class="btn btn-success" style="margin-bottom:8px;"> Take Snapshot and Identify</button>
-                <button id="btn-start" type="button" class="btn" style="margin-bottom:8px;">Start</button>
-        <button id="btn-stop" type="button" class="btn" style="margin-bottom:8px;">Stop</button>
-                </div>
-            </div>
-        </div>*@
-
-        @*<div class="col-md-6">
-            <div class="panel panel-primary">
-                <div class="panel-heading">Captured Photo</div>
-                <div class="panel-body" style="text-align:center">
-                    <canvas id="capture" width="320" height="240"></canvas>
-                    <div id="snapshot">
-                        @if (ViewBag.imgcap != null)
-                        {
-                            <input type="image" disabled="disabled" src="@ViewBag.imgcap"/>
-                        }
-                        else
-                        {
-                        <input type="image" disabled="disabled" id="imgcap" name="imgcap" style="width:320px;height:240px;display:none" />
-                        }
-                        <input type="text" id="txtimgcap" style="display:none" asp-for="CapturedFile" />
-                    </div>
-                    <br />
-                    <br />
-                    <br />
-                </div>
-            </div>
-        </div>*@
-        @*@if (ViewBag.Result != null)
-        {
-            <div class="col-md-12">
-                <div class="panel panel-primary">
-                    <div class="bg-success createperson" id="divResult" style="text-align:center">
-                        @ViewBag.Result
-                    </div>
-                </div>
-            </div>
-        }*@
-    </form>
+<div class="text-center">
+    <h1 class="display-4">Welcome</h1>
+    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
 </div>
-<script src="~/lib/jquery/dist/jquery.js"></script>
-    @*<script src="~/lib/jquery/dist/Webcam.js"></script>*@
-<script language="javaScript">
-
-    // The buttons to start & stop stream and to capture the image
-    //var btnStart = document.getElementById("btn-start");
-    //var btnStop = document.getElementById("btn-stop");
-    var btnCapture = $("#btn-capture")[0];
-    // The stream & capture
-    var stream = document.getElementById("stream");
-    var capture = document.getElementById("capture");
-    var snapshot = document.getElementById("snapshot");
-
-    // The video stream
-    var cameraStream = null;
-
-    // Attach listeners
-    //btnStart.addEventListener("click", startStreaming);
-    //btnStop.addEventListener("click", stopStreaming);
-    if (typeof btnCapture !== 'undefined') {
-        btnCapture.addEventListener("click", captureSnapshot);
-    }
-    var loadType = '';
-    $(document).ready(function () {
-        loadType = '@ViewBag.imgcap';        
-        
-        if (loadType == '') {
-            $(window).on("load", startStreaming());
-        }
-        else {
-            $(window).on("load", stopStreaming());            
-        }
-    });
-
-
-    
-    
-    $(window).on('beforeunload', function () {
-            stopStreaming();
-        });
-
-        ////imgInp.onchange = evt => {
-        ////    const [file] = imgInp.files
-        ////    if (file) {
-        ////        imgdisp.src = URL.createObjectURL(file)
-        ////    }
-        ////}
-        // Start Streaming
-
-        function startStreaming() {
-            var constraints;
-            var isFirefox = typeof InstallTrigger !== 'undefined';            
-
-            $("#browerName").val(isFirefox);
-
-            var mediaSupport = 'mediaDevices' in navigator;
-
-            if (mediaSupport && null == cameraStream) {
-
-                if (isFirefox) {
-                    constraints = {
-                        audio: false,
-                    video: {
-                            width: { max: 352 },
-                        height: {max: 378 }
-                        },
-                };
-                }
-                else {                    
-                    constraints = {
-                        audio: false,
-                    video: {
-                            width: { min: 338, max: 338 },
-                        height: { min: 378, max: 378 }
-                        },
-                };
-                }
-
-                navigator.mediaDevices.getUserMedia(constraints)
-                    .then(function (mediaStream) {
-                    cameraStream = mediaStream;
-                    stream.srcObject = mediaStream;
-                    stream.play();
-                })
-                .catch(function (err) {
-
-            console.log("Unable to access camera: " + err);
-        });
-        }
-        else {
-
-            alert('Your browser does not support media devices.');
-
-            return;
-        }
-    }
-
-    // Stop Streaming
-    function stopStreaming() {
-
-        if (null != cameraStream) {
-
-            var track = cameraStream.getTracks()[0];
-
-            track.stop();
-            stream.load();
-
-            cameraStream = null;
-        }
-    }
-
-    function captureSnapshot() {        
-        if (null != cameraStream) {
-            var isFirefox = typeof InstallTrigger !== 'undefined';
-            divresult=document.getElementById('divResult')
-            if (divresult!= undefined) {
-                divresult.innerHTML = "";
-                divresult.style.display = 'none';
-            }           
-            var ctx = capture.getContext('2d');
-            var img = document.getElementById("imgcap");
-            img.style.display = 'block';
-            ctx.drawImage(stream, 0, 0, capture.width, capture.height);
-
-            img.src = capture.toDataURL("image/png");
-            img.width = 338;
-            img.height = 378;
-            var txt = document.getElementById("txtimgcap")           
-            if (isFirefox) {
-                img.style.display = 'none';
-            }
-            txt.value = img.src;
-            
-        }
-    }
-</script>
