@@ -1,3 +1,3 @@
-@using FaceApp
+﻿@using FaceApp
 @using FaceApp.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
